refactor front end
