+++ client/index.html
@@ -10,7 +10,7 @@
 <body id="body">
     <div id="container">
         <h1 id="title">TV Show Quotes</h1>
-        <button id="toogle-dark-mode" class="border-4 border-light-blue-500 border-solid">dark mode</button>
+        <button id="toogle-dark-mode">dark mode</button>
 
         <div id="image-wrapper">
             <img id ="background_image" src='/assets/movie.jpeg' alt="restaurant_image">
@@ -20,6 +20,6 @@
         <p id="add-success"></p>
 
     </div>
-    <script src='/src/index.js'></script>
+    <script src='/src/index.js' type="module"></script>
 </body>
 </html>
